merge ad 2020-04-02

+++ redisson/src/test/java/org/redisson/BaseTest.java
@@ -1,16 +1,15 @@
 package org.redisson;
 
 import java.io.IOException;
+
 import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.redisson.api.RedissonClient;
-import org.redisson.codec.MsgPackJacksonCodec;
 import org.redisson.config.Config;
 
 public abstract class BaseTest {
-
+    
     protected RedissonClient redisson;
     protected static RedissonClient defaultRedisson;
 
@@ -19,14 +18,12 @@ public abstract class BaseTest {
         if (!RedissonRuntimeEnvironment.isTravis) {
             RedisRunner.startDefaultRedisServerInstance();
             defaultRedisson = createInstance();
-        }
-    }
-
-    @AfterClass
-    public static void afterClass() throws IOException, InterruptedException {
-        if (!RedissonRuntimeEnvironment.isTravis) {
-            defaultRedisson.shutdown();
-            RedisRunner.shutDownDefaultRedisServerInstance();
+            Runtime.getRuntime().addShutdownHook(new Thread() {
+                @Override
+                public void run() {
+                    defaultRedisson.shutdown();
+                }
+            });
         }
     }
 
@@ -39,7 +36,9 @@ public abstract class BaseTest {
             if (redisson == null) {
                 redisson = defaultRedisson;
             }
-            redisson.getKeys().flushall();
+            if (flushBetweenTests()) {
+                redisson.getKeys().flushall();
+            }
         }
     }
 
@@ -52,15 +51,16 @@ public abstract class BaseTest {
     }
 
     public static Config createConfig() {
-        String redisAddress = System.getProperty("redisAddress");
-        if (redisAddress == null) {
-            redisAddress = "127.0.0.1:6379";
-        }
+//        String redisAddress = System.getProperty("redisAddress");
+//        if (redisAddress == null) {
+//            redisAddress = "127.0.0.1:6379";
+//        }
         Config config = new Config();
 //        config.setCodec(new MsgPackJacksonCodec());
 //        config.useSentinelServers().setMasterName("mymaster").addSentinelAddress("127.0.0.1:26379", "127.0.0.1:26389");
 //        config.useClusterServers().addNodeAddress("127.0.0.1:7004", "127.0.0.1:7001", "127.0.0.1:7000");
-        config.useSingleServer().setAddress(redisAddress);
+        config.useSingleServer()
+                .setAddress(RedisRunner.getDefaultRedisServerBindAddressAndPort());
 //        .setPassword("mypass1");
 //        config.useMasterSlaveConnection()
 //        .setMasterAddress("127.0.0.1:6379")
@@ -74,4 +74,7 @@ public abstract class BaseTest {
         return Redisson.create(config);
     }
 
+    protected boolean flushBetweenTests() {
+        return true;
+    }
 }
